requirements.txt
fastapi==0.111.0
uvicorn==0.23.0
openai==1.34.0
python-dotenv==1.1.1
twilio
httpx
